requirements.txt
pandas
matplotlib
python-dotenv
numpy
requests
fastapi
uvicorn